+++ didattica/js/sportello.js
@@ -6,6 +6,8 @@
  */
 
 var soloNuovi=1;
+var categoria_filtro_id=0;
+var docente_filtro_id=0;
 var materia_filtro_id=0;
 
 function setDbDateToPickr(pickr, data_str) {
@@ -30,7 +32,7 @@ $('#soloNuoviCheckBox').change(function() {
 });
 
 function sportelloReadRecords() {
-	$.get("sportelloReadRecords.php?ancheCancellati=true&soloNuovi=" + soloNuovi + "&materia_filtro_id=" + materia_filtro_id, {}, function (data, status) {
+	$.get("sportelloReadRecords.php?ancheCancellati=true&soloNuovi=" + soloNuovi + "&categoria_filtro_id=" + categoria_filtro_id + "&docente_filtro_id=" + docente_filtro_id + "&materia_filtro_id=" + materia_filtro_id, {}, function (data, status) {
 		$(".records_content").html(data);
         $('[data-toggle="tooltip"]').tooltip({
             container: 'body'
@@ -54,6 +56,22 @@ function sportelloDelete(id, materia) {
 }
 
 function sportelloSave() {
+	if ($("#materia").val() <= 0) {
+		$("#_error-materia").text("Devi selezionare una materia");
+		$("#_error-materia-part").show();
+		return;
+	}
+	if ($("#docente").val() <= 0) {
+		$("#_error-materia").text("Devi selezionare un docente");
+		$("#_error-materia-part").show();
+		return;
+	}
+	if ($("#numero_ore").val() <= 0) {
+		$("#_error-materia").text("Il numero di ore non può essere 0");
+		$("#_error-materia-part").show();
+		return;
+	}
+    $("#_error-materia-part").hide();
     $.post("sportelloSave.php", {
         id: $("#hidden_sportello_id").val(),
 		data: getDbDateFromPickrId("#data"),
@@ -62,10 +80,12 @@ function sportelloSave() {
 		materia_id: $("#materia").val(),
         numero_ore: $("#numero_ore").val(),
 		argomento: $("#argomento").val(),
-		luogo: $("#commento").val(),
+		luogo: $("#luogo").val(),
+        max_iscrizioni: $("#max_iscrizioni").val(),
         classe: $("#classe").val(),
-        cancellato: $("#cancellato").prop('checked'),
-        firmato: $("#firmato").prop('checked')
+        cancellato: $("#cancellato").is(':checked')? 1: 0,
+        firmato: $("#firmato").is(':checked')? 1: 0,
+        online: $("#online").is(':checked')? 1: 0
     }, function (data, status) {
         $("#sportello_modal").modal("hide");
         sportelloReadRecords();
@@ -88,8 +108,10 @@ function sportelloGetDetails(sportello_id) {
             $("#argomento").val(sportello.sportello_argomento);
             $("#luogo").val(sportello.sportello_luogo);
             $("#classe").val(sportello.sportello_classe);
+            $("#max_iscrizioni").val(sportello.sportello_max_iscrizioni);
             $("#cancellato").prop('checked', sportello.sportello_cancellato != 0 && sportello.sportello_cancellato != null);
             $("#firmato").prop('checked', sportello.sportello_firmato != 0 && sportello.sportello_firmato != null);
+            $("#online").prop('checked', sportello.sportello_online != 0 && sportello.sportello_online != null);
         });
     } else {
         data_pickr.setDate(Date.today().toString('d/M/yyyy'));
@@ -102,11 +124,29 @@ function sportelloGetDetails(sportello_id) {
         $("#argomento").val("");
         $("#luogo").val("");
         $("#classe").val("");
+        $("#max_iscrizioni").val($("#hidden_max_iscrizioni_default").val());
         $("#cancellato").prop('checked', false);
         $("#firmato").prop('checked', false);
+        $("#onine").prop('checked', false);
+    }
+	$("#_error-materia-part").hide();
+	$("#sportello_modal").modal("show");
 }
 
-	$("#sportello_modal").modal("show");
+function importFile(file) {
+    var contenuto = "";
+    const reader = new FileReader();
+    reader.addEventListener('load', (event) => {
+        contenuto = event.target.result;
+        $.post("sportelloImport.php", {
+            contenuto: contenuto
+        },
+        function (data, status) {
+            $('#result_text').html(data);
+            sportelloReadRecords();
+        });
+    });
+    reader.readAsText(file);
 }
 
 $(document).ready(function () {
@@ -119,10 +159,25 @@ $(document).ready(function () {
 
     sportelloReadRecords();
     
+    $("#categoria_filtro").on("changed.bs.select", 
+    function(e, clickedIndex, newValue, oldValue) {
+        categoria_filtro_id = this.value;
+        sportelloReadRecords();
+    });
+    
+    $("#docente_filtro").on("changed.bs.select", 
+    function(e, clickedIndex, newValue, oldValue) {
+        docente_filtro_id = this.value;
+        sportelloReadRecords();
+    });
+    
     $("#materia_filtro").on("changed.bs.select", 
     function(e, clickedIndex, newValue, oldValue) {
         materia_filtro_id = this.value;
         sportelloReadRecords();
     });
 
+    $('#file_select_id').change(function (e) {
+        importFile(e. target. files[0]);
+    });
 });