+++ didattica/sportelloSave.php
@@ -10,8 +10,7 @@
  require_once '../common/checkSession.php';
  ruoloRichiesto('segreteria-didattica');
 
- $tableName = "sportello";
- if(isset($_POST)) {
+if(isset($_POST)) {
 	$id = $_POST['id'];
 	$data = $_POST['data'];
 	$ora = $_POST['ora'];
@@ -21,18 +20,20 @@
 	$argomento = escapePost('argomento');
 	$luogo = escapePost('luogo');
 	$classe = escapePost('classe');
+	$max_iscrizioni = $_POST['max_iscrizioni'];
 	$cancellato = $_POST['cancellato'];
 	$firmato = $_POST['firmato'];
+	$online = $_POST['online'];
 
 	if ($id > 0) {
-        $query = "UPDATE $tableName SET data = '$data', ora = '$ora', docente_id = '$docente_id', materia_id = '$materia_id', numero_ore = '$numero_ore', argomento = '$argomento', luogo = '$luogo', classe = '$classe', cancellato = $cancellato, firmato = $firmato WHERE id = '$id'";
-        dbExec($query);
-        info("aggiornato $tableName id=$id data=$data ora=$ora docente_id=$docente_id materia_id=$materia_id numero_ore=$numero_ore argomento=$argomento luogo=$luogo classe=$classe");
-    } else {
-        $query = "INSERT INTO $tableName(data, ora, docente_id, materia_id, numero_ore, argomento, luogo, classe, anno_scolastico_id) VALUES('$data', '$ora', '$docente_id', '$materia_id', '$numero_ore', '$argomento', '$luogo', '$classe', $__anno_scolastico_corrente_id)";
-        dbExec($query);
-        $id = dblastId();
-		info("aggiunto sportello id=$id data=$data ora=$ora docente_id=$docente_id materia_id=$materia_id numero_ore=$numero_ore argomento=$argomento luogo=$luogo classe=$classe");
-    }
+		$query = "UPDATE sportello SET data = '$data', ora = '$ora', docente_id = '$docente_id', materia_id = '$materia_id', numero_ore = '$numero_ore', argomento = '$argomento', luogo = '$luogo', classe = '$classe', max_iscrizioni = '$max_iscrizioni', cancellato = $cancellato, firmato = $firmato, online = $online WHERE id = '$id'";
+		dbExec($query);
+		info("aggiornato sportello id=$id data=$data ora=$ora docente_id=$docente_id materia_id=$materia_id numero_ore=$numero_ore argomento=$argomento luogo=$luogo classe=$classe max_iscrizioni=$max_iscrizioni online = $online");
+	} else {
+		$query = "INSERT INTO sportello(data, ora, docente_id, materia_id, numero_ore, argomento, luogo, classe, max_iscrizioni, online, anno_scolastico_id) VALUES('$data', '$ora', '$docente_id', '$materia_id', '$numero_ore', '$argomento', '$luogo', '$classe', '$max_iscrizioni', '$online', $__anno_scolastico_corrente_id)";
+		dbExec($query);
+		$id = dblastId();
+		info("aggiunto sportello id=$id data=$data ora=$ora docente_id=$docente_id materia_id=$materia_id numero_ore=$numero_ore argomento=$argomento luogo=$luogo classe=$classe max_iscrizioni=$max_iscrizioni online = $online");
+	}
 }
 ?>