+++ changelog.md
@@ -1,3 +1,19 @@
+## Version 1.2.8 - 10 ott 2021
+
+##### Improvements
+- sportello import
+- template xlsx per import sportello
+- categoria sportello
+- aggiornato sql
+
+## Version 1.2.7 - 4 ott 2021
+
+##### Improvements
+- sportello online (flag)
+- studente login separato (non utente)
+- sportelloCheck per controllare se ci sono iscritti (incluso invio email)
+- aggiunta cartella template per i modelli per import di docenti, studenti, gruppi e corsi di recupero
+
 ## Version 1.2.6 - 25 set 2021
 
 ##### Improvements